added printptr, + added long unsigned hex to prevent hex conversion errors

+++ ft_printf/ft_printhexupp.c
@@ -1,13 +1,12 @@
 #include "libftprintf.h"
-#include <stdio.h>
 
 int	ft_printhexupp(va_list args)
 {
-	int arg;
+	long unsigned int arg;
 	int index;
 	char * hex;
 	
-	arg = va_arg(args, int);
+	arg = va_arg(args, long unsigned int);
 	hex = ft_hexconvert(arg);
 	index = ft_strlen(hex); 
 	while (index >= 0)
